requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.1
google-cloud-firestore==2.16.1
